wsgi.py

from os import environ, path
from dotenv import load_dotenv

from panel import create_app

basedir = path.abspath(path.dirname(__file__))
load_dotenv(path.join(basedir, ".env"))

app = create_app(environ.get("FLASK_ENV", "development"))

if __name__ == "__main__":
    app.run(host="0.0.0.0")
