pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Project zomboid server panel"
authors = ["Emilio <emilio2hd@gmail.com>"]
readme = "README.md"
repository = "https://github.com/emilio2hd/pz-panel"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.3"
Flask-Login = "^0.5.0"
psutil = "^5.9.0"
rcon = "1.3.7"
transitions = "^0.8.11"
gunicorn = "^20.1.0"
Flask-Assets = "^2.0"
jsmin = "^3.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
python-dotenv = "^0.19.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
